# api/openapi.yaml
openapi: 3.1.0
info:
  title: base-ecommerce API
  version: 0.10.0
  description: |
    Hợp đồng API. Đây là NGUỒN SỰ THẬT — sửa endpoint thì sửa file này trước.
servers:
  - url: http://localhost:8080/api/v1
    description: Máy dev
paths:
  /products:
    get:
      operationId: listProducts
      summary: Danh sách sản phẩm có lọc và phân trang
      security: []   # công khai — tuyên bố chủ động, không suy ra từ việc thiếu
      description: |
        Ngoài các tham số liệt kê dưới đây, endpoint còn chấp nhận lọc theo
        thuộc tính tự do qua tiền tố `attr.`: mỗi cặp `attr.<tên>=<giá trị>`
        lọc sản phẩm có thuộc tính `<tên>` bằng đúng `<giá trị>`, ví dụ
        `?attr.ram=16GB&attr.socket=AM5`. `<tên>` là tên thuộc tính bất kỳ do
        người bán tự đặt (lưu trong `Product.attributes`), nên không thể liệt
        kê hết trong OpenAPI — không có giới hạn nào ngoài quy ước tiền tố này.

        ⚠️ Type sinh bằng `openapi-typescript` KHÔNG phủ tham số này: object
        `query` sinh ra là object đóng, không có index signature. Frontend phải
        tự nối `attr.<tên>=<giá trị>` vào query string thay vì truyền qua tham
        số `query` đã sinh type.
      parameters:
        # Mọi description đều phải nằm trong nháy kép. Trong flow mapping { }
        # dấu phẩy KẾT THÚC plain scalar, nên `description: Giá tối thiểu, chuỗi
        # thập phân VND.` bị parse thành description "Giá tối thiểu" CỘNG một
        # property rác tên "chuỗi thập phân VND.". File vẫn sinh được type TS nên
        # openapi-drift xanh, nhưng validator OpenAPI 3.1 báo lỗi.
        - name: category
          in: query
          schema: { type: string }
          description: "Lọc theo slug danh mục, bao gồm toàn bộ danh mục con."
        - name: brand
          in: query
          schema: { type: string }
          description: "Lọc theo slug thương hiệu."
        - name: price_min
          in: query
          schema: { type: string }
          description: "Giá tối thiểu, chuỗi thập phân VND."
        - name: price_max
          in: query
          schema: { type: string }
          description: "Giá tối đa, chuỗi thập phân VND."
        - name: sort
          in: query
          schema: { type: string, enum: [newest, price_asc, price_desc] }
          description: "Giá trị ngoài danh sách này trả 400 INVALID_SORT."
        - name: page
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 1 }
          description: >-
            Vượt 200 trả 400 PAGE_TOO_DEEP: offset sâu quá đắt. Không phải số
            trả 400 INVALID_PAGINATION. Nhỏ hơn 1 thì bị kẹp về 1.
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 24 }
          description: >-
            Không phải số trả 400 INVALID_PAGINATION. Ngoài khoảng thì bị KẸP về
            biên chứ không bị từ chối: 0 và số âm thành 24, lớn hơn 100 thành
            100. Đọc `meta.limit` để biết server thật sự dùng bao nhiêu.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProductList" }
        "400": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        "500": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
...
  /categories:
    get:
      operationId: getCategoryTree
      security: []
      summary: Cây danh mục
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Category" }
        "500": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
  /brands:
    get:
      operationId: listBrands
      security: []
      summary: Toàn bộ thương hiệu, sắp theo tên
      description: >-
        Danh sách ngắn (vài trăm), không phân trang. Frontend dùng để hiện tên
        thương hiệu từ `Product.brand_id`.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Brand" }
        "500": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
  /sitemap/products:
    get:
      operationId: sitemapProducts
      security: []
      summary: Slug + ngày cập nhật của sản phẩm live, cho sitemap phân mảnh
      description: >-
        5.000 dòng mỗi trang, sắp theo id — sản phẩm mới luôn rơi vào trang
        CUỐI, trang cũ không bị dồn dịch. Không có trần 200 trang như GET
        /products (câu này chạy trên index hẹp, Index Only Scan). Trang vượt
        quá trả `data` rỗng, không báo lỗi.
      parameters:
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      required: [slug, updated_at]
                      properties:
                        slug:       { type: string }
                        updated_at: { type: string, format: date-time }
                  meta:
                    type: object
                    required: [page, page_size, total, total_pages]
                    properties:
                      page:        { type: integer }
                      page_size:   { type: integer }
                      total:       { type: integer }
                      total_pages: { type: integer }
        "400": { $ref: "#/components/responses/Problem" }
        "500": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
...
  /attributes:
    get:
      operationId: listAttributes
      security: []
      summary: Mọi định nghĩa thuộc tính
      description: Danh sách ngắn, không phân trang. Storefront dùng để đổi mã thuộc tính ra tên và đơn vị.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Attribute" }
        "500": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
...
  /admin/categories:
    post:
      operationId: createCategory
      summary: Tạo danh mục
      security: [{ bearerAuth: [] }]
      x-required-permission: catalog.taxonomy.write
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateCategoryRequest" }
      responses:
        "201":
          description: Đã tạo
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CategoryNode" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "413": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        "500": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
  /admin/categories/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    patch:
      operationId: updateCategory
      summary: Sửa, đổi tên, chuyển danh mục
      security: [{ bearerAuth: [] }]
      x-required-permission: catalog.taxonomy.write
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateCategoryRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CategoryNode" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "413": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        "500": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
    delete:
      operationId: deleteCategory
      summary: Xóa danh mục rỗng
      description: >-
        Chỉ xóa được danh mục không còn danh mục con (409 CATEGORY_HAS_CHILDREN)
        và không còn sản phẩm nào, KỂ CẢ sản phẩm đã xóa mềm (409
        CATEGORY_HAS_PRODUCTS).
      security: [{ bearerAuth: [] }]
      x-required-permission: catalog.taxonomy.write
      responses:
        "204": { description: Đã xóa }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "500": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
  /admin/brands:
    post:
      operationId: createBrand
      summary: Tạo thương hiệu
      security: [{ bearerAuth: [] }]
      x-required-permission: catalog.taxonomy.write
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateBrandRequest" }
      responses:
        "201":
          description: Đã tạo
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Brand" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "413": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        "500": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
  /admin/brands/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    patch:
      operationId: updateBrand
      summary: Sửa thương hiệu
      security: [{ bearerAuth: [] }]
      x-required-permission: catalog.taxonomy.write
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateBrandRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Brand" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "413": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        "500": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
    delete:
      operationId: deleteBrand
      summary: Xóa thương hiệu không còn sản phẩm
      security: [{ bearerAuth: [] }]
      x-required-permission: catalog.taxonomy.write
      responses:
        "204": { description: Đã xóa }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "500": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
  /admin/products:
    post:
      operationId: createProduct
      summary: Tạo sản phẩm
      security: [{ bearerAuth: [] }]
      x-required-permission: catalog.products.write
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateProductRequest" }
      responses:
        "201":
          description: Đã tạo
          headers:
            Location: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "413": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        "500": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
  /admin/products/{id}:
    patch:
      operationId: updateProduct
      summary: Sửa sản phẩm
      security: [{ bearerAuth: [] }]
      x-required-permission: catalog.products.write
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateProductRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        # 409 rất hay gặp ở màn hình sửa: đổi tên sinh slug mới, đụng slug của
        # sản phẩm khác -> DUPLICATE_SLUG.
        "409": { $ref: "#/components/responses/Problem" }
        "413": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        "500": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
  /admin/products/{id}/variants:
    post:
      operationId: addVariant
      summary: Thêm phiên bản cho sản phẩm
      description: Trả về CẢ sản phẩm — giá "từ" của sản phẩm có thể đổi theo.
      security: [{ bearerAuth: [] }]
      x-required-permission: catalog.products.write
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateVariantRequest" }
      responses:
        "201":
          description: Đã thêm
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "413": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        "500": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
  /admin/products/{id}/variants/{variantId}:
    patch:
      operationId: updateVariant
      summary: Sửa phiên bản (giá, tùy chọn, trạng thái, thứ tự)
      description: >-
        Không có DELETE: đơn hàng sẽ trỏ vào phiên bản, nên ngừng bán là
        `status: inactive`. Trả về CẢ sản phẩm.
      security: [{ bearerAuth: [] }]
      x-required-permission: catalog.products.write
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: variantId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateVariantRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "413": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        "500": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
  /admin/products/{id}/publish:
    post:
      operationId: publishProduct
      summary: Đăng bán sản phẩm
      security: [{ bearerAuth: [] }]
      x-required-permission: catalog.products.write
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        "500": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        Access token từ /auth/login, sống 15 phút. Chỉ mang định danh người
        dùng — quyền được tra mỗi request.
  responses:
    Problem:
      description: Lỗi theo RFC 7807
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
...
    Product:
      type: object
      required: [id, slug, name, short_description, category_id, brand_id,
                 price, currency, status, attributes, images, variants, created_at, updated_at]
      properties:
        id:                { type: string, format: uuid }
        slug:              { type: string }
        name:              { type: string }
        short_description: { type: string }
        category_id:       { type: string, format: uuid }
        brand_id:          { type: string, format: uuid }
        price:
          type: string
          description: |
            GIÁ "TỪ": giá thấp nhất trong các phiên bản đang bán (từ 0.4.0; trước
            đó là giá của sản phẩm duy nhất). Lọc `price_min`/`price_max` và sắp
            xếp theo giá đều dùng con số này. Chuỗi thập phân. Số 0 thừa ở cuối BỊ CẮT ("25990000.50" đọc lại thành "25990000.5"), nên đừng so sánh bằng chuỗi — hãy parse bằng thư viện decimal. KHÔNG BAO GIỜ dùng number:
            JSON number qua JS mất chính xác với NUMERIC(15,2) lớn (IEEE 754
            double chỉ biểu diễn chính xác số nguyên tới 2^53), nên giá đọc lại
            có thể lệch so với giá đã ghi. Server luôn trả amount đã được
            round-trip qua shopspring/decimal, không định dạng hiển thị (dấu
            phẩy nghìn, ký hiệu tiền tệ...) — việc đó là của frontend.
        currency:
          type: string
          description: Hiện chỉ có "VND" — hệ thống chưa hỗ trợ đa tiền tệ.
          examples: ["VND"]
        status:     { type: string, enum: [draft, live, archived] }
        attributes: { type: object, additionalProperties: { type: string } }
        images:
          type: array
          items: { type: string }
          description: >-
            KHÓA media (`Media.key`, ví dụ "products/0190….jpg"), KHÔNG phải URL
            — từ 0.6.0. Hiển thị qua imgproxy của storefront. Ghi sản phẩm với
            key chưa ready trả 422, trường images[i] mã UNKNOWN_IMAGE.
        variants:
          type: array
          description: >-
            Sắp theo `position` rồi `id`. Endpoint công khai chỉ trả phiên bản
            `active`; response của các lệnh quản trị trả đủ, kể cả `inactive`.
          items: { $ref: "#/components/schemas/Variant" }
        created_at:
          type: string
          format: date-time
          description: RFC 3339, luôn ở múi giờ UTC (hậu tố "Z").
        updated_at:
          type: string
          format: date-time
          description: RFC 3339, luôn ở múi giờ UTC (hậu tố "Z").
    ProductList:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/Product" }
        meta:
          type: object
          required: [page, limit, total, total_pages, max_page, has_next, has_prev]
          properties:
            page:        { type: integer }
            limit:       { type: integer }
            total:       { type: integer }
            total_pages: { type: integer }
            max_page:
              type: integer
              description: >-
                Trang sâu nhất truy cập được (vượt qua trả 400 PAGE_TOO_DEEP).
                `has_next` đã tính tới trần này, nên cứ theo `has_next` là an
                toàn: nó về false ở trang `max_page` dù `total_pages` còn lớn hơn.
            has_next:    { type: boolean }
            has_prev:    { type: boolean }
    Category:
      type: object
      required: [id, slug, name, position, children]
      properties:
        id:       { type: string, format: uuid }
        slug:     { type: string }
        name:     { type: string }
        position: { type: integer }
        children:
          type: array
          items: { $ref: "#/components/schemas/Category" }
...
    CreateVariantRequest:
      type: object
      additionalProperties: false
      required: [sku, price]
      properties:
        sku:      { type: string, maxLength: 64 }
        price:    { type: string, description: "Chuỗi thập phân VND. Nháp chưa có giá thì gửi \"0\"." }
        currency: { type: string, default: "VND" }
        options:
          type: object
          additionalProperties: { type: string }
          description: "Tối đa 10 cặp. Không được trùng options của phiên bản khác cùng sản phẩm (409 DUPLICATE_VARIANT_OPTIONS)."
        position: { type: integer, default: 0 }
...
    Attribute:
      type: object
      required: [id, code, name, type, unit, options, filterable, variant]
      properties:
        id:   { type: string, format: uuid }
        code: { type: string, description: "Khóa trong attributes/options và trong tham số attr.<code>. Bất biến." }
        name: { type: string }
        type:
          type: string
          enum: [text, number, boolean, enum]
          description: >-
            Bất biến. number: số thập phân thuần ("16", "15.6") — đơn vị nằm ở
            `unit`, không nằm trong giá trị. boolean: "true"/"false".
        unit:       { type: string }
        options:    { type: array, items: { type: string }, description: "Giá trị cho phép của kiểu enum; rỗng với kiểu khác." }
        filterable: { type: boolean, description: "Có hiện trong /products/facets không." }
        variant:
          type: boolean
          description: >-
            Bất biến. true: chỉ đặt trong Variant.options (RAM, màu). false: chỉ
            đặt trong Product.attributes (CPU). Đặt sai chỗ trả 422.
...
    Brand:
      type: object
      required: [id, slug, name]
      properties:
        id:   { type: string, format: uuid }
        slug: { type: string }
        name: { type: string }
    CategoryNode:
      type: object
      description: Một danh mục đứng riêng, không kèm cây con — trả về từ API quản trị.
      required: [id, parent_id, slug, name, position]
      properties:
        id:        { type: string, format: uuid }
        parent_id: { type: ["string", "null"], format: uuid, description: "null nghĩa là danh mục gốc." }
        slug:      { type: string }
        name:      { type: string }
        position:  { type: integer }
    CreateCategoryRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name:      { type: string, maxLength: 100 }
        slug:
          type: string
          description: "Bỏ trống thì sinh từ tên. Có thì vẫn được chuẩn hóa (bỏ dấu, chữ thường)."
        parent_id: { type: ["string", "null"], format: uuid, description: "Bỏ trống hoặc null là danh mục gốc." }
        position:  { type: integer, default: 0 }
    UpdateCategoryRequest:
      type: object
      additionalProperties: false
      description: >-
        Sửa TỪNG PHẦN, trường vắng mặt giữ nguyên. Đổi `name` KHÔNG đổi `slug`
        (slug nằm trong URL đã được index) — muốn đổi slug thì gửi `slug`.
      properties:
        name:      { type: string, maxLength: 100 }
        slug:      { type: string }
        parent_id:
          type: ["string", "null"]
          format: uuid
          description: >-
            BA trạng thái: vắng mặt = giữ nguyên chỗ; null = lên làm danh mục
            gốc; uuid = chuyển vào dưới danh mục đó. Chuyển vào chính nó hoặc
            con cháu của nó trả 422 CATEGORY_CYCLE.
        position:  { type: integer }
    CreateBrandRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name: { type: string, maxLength: 100 }
        slug: { type: string, description: "Bỏ trống thì sinh từ tên." }
    UpdateBrandRequest:
      type: object
      additionalProperties: false
      description: "Sửa từng phần. Đổi `name` KHÔNG đổi `slug`."
      properties:
        name: { type: string, maxLength: 100 }
        slug: { type: string }
    CreateProductRequest:
      type: object
      # Server dùng DisallowUnknownFields: trường lạ trả 400 MALFORMED_REQUEST.
      # Thiếu dòng này thì theo JSON Schema nghĩa là CHO PHÉP trường thêm, và
      # frontend gửi kèm một field của UI sẽ ăn 400 dù spec bảo hợp lệ.
      additionalProperties: false
      required: [name, category_id, brand_id, variants]
      properties:
        name:              { type: string, maxLength: 200 }
        short_description: { type: string }
        category_id:       { type: string, format: uuid }
        brand_id:          { type: string, format: uuid }
        variants:
          type: array
          minItems: 1
          description: >-
            Ít nhất một phiên bản (422 VARIANT_REQUIRED). Từ 0.4.0 `sku` và
            `price` nằm ở đây, không còn ở cấp sản phẩm.
          items: { $ref: "#/components/schemas/CreateVariantRequest" }
        attributes:        { type: object, additionalProperties: { type: string } }
        images:
          type: array
          items: { type: string }
          description: >-
            KHÓA media (`Media.key`, ví dụ "products/0190….jpg"), KHÔNG phải URL
            — từ 0.6.0. Hiển thị qua imgproxy của storefront. Ghi sản phẩm với
            key chưa ready trả 422, trường images[i] mã UNKNOWN_IMAGE.
    UpdateProductRequest:
      type: object
      additionalProperties: false
      description: >-
        Sửa TỪNG PHẦN: mọi trường đều không bắt buộc, và trường vắng mặt GIỮ
        NGUYÊN giá trị đang có. Gửi `{}` là lệnh không đổi gì.


        Ràng buộc của sản phẩm đang "live" được kiểm trên trạng thái SAU khi
        sửa: không thể PATCH để nó còn 0 ảnh (422 NO_IMAGE).


        Từ 0.4.0 không còn `price`/`currency`: giá thuộc phiên bản — sửa qua
        PATCH /admin/products/{id}/variants/{variantId}. Gửi `price` ở đây trả
        400 MALFORMED_REQUEST.
      properties:
        name:
          type: string
          maxLength: 200
          description: "Đổi tên sẽ sinh lại slug; trùng slug sản phẩm khác trả 409 DUPLICATE_SLUG."
        short_description: { type: string }
        attributes:        { type: object, additionalProperties: { type: string } }
        images:
          type: array
          items: { type: string }
          description: >-
            KHÓA media (`Media.key`, ví dụ "products/0190….jpg"), KHÔNG phải URL
            — từ 0.6.0. Hiển thị qua imgproxy của storefront. Ghi sản phẩm với
            key chưa ready trả 422, trường images[i] mã UNKNOWN_IMAGE.
        category_id:
          type: string
          format: uuid
          description: "Chuyển sang danh mục khác. Không tồn tại trả 422 CATEGORY_NOT_FOUND."
        brand_id:
          type: string
          format: uuid
          description: "Đổi thương hiệu. Không tồn tại trả 422 BRAND_NOT_FOUND."